Use correct nxi value for resampled fields
When increasing the number of grid points by 2 in each direction, the
value of nxi is also multiplied by 2.

Before, the values of the momentum, and thus the circulation, were
incorrect when working with resampled fields. For example, for
resampling_factor = 2, the circulation of the velocity took values
multiple of κ/2 instead of κ.

This also modifies the loop_sizes values saved to the output HDF5 files.
The values now refer to the original field and not to the resampled one.