Add option to compute velocity increment stats
This needs to be enabled in the parameter file, see
examples/tangle.toml.

The parameters for the increments are in a new "[increments]" section.
Parameters are almost the same as for the circulation, with small
changes like "loop_sizes" -> "spatial_offsets" and "max_kappa" -> "max_c".

In both the circulation and increment sections, there is a new "enabled"
parameter. Its value must be "true" for exactly one of the two stats.